moved validator code to utils.ts
